feat(chess-engine): update is\_square\_attacked function to accept PIECE\_COLOR parameter instead of piece\_color
The change was made to improve the consistency and readability of the code by using a constant variable name PIECE\_COLOR for the boolean parameter that represents the color of the attacking pieces. This makes it clearer that the same parameter is used throughout the function, rather than having a different variable name in some places. Additionally, this change allows for easier future modifications if needed, as the code will be more self-documenting and easier to understand.