feat: simplify .gitignore (CLion template)
Remove unnecessary lines from the CLion template in .gitignore, reducing its size and improving readability. This change also removes duplicate entries for C++ files.

feat(README.md): add build & setup instructions and prerequisites
Adds a `setup.sh` script to configure the project, builds the engine and tests, and installs a pre-commit Git hook (if this is a Git repository). The README now includes instructions on how to build the project in both Debug and Release modes, as well as prerequisites for building the project. This commit also adds a new section for common commands when using the Debug build.